requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
python-multipart>=0.0.12
jinja2>=3.1
beautifulsoup4>=4.12
lxml>=5.3
pypdf>=5.1
pydantic>=2.9
google-genai>=0.3
python-dotenv>=1.0
httpx>=0.27
pytest>=8.3
